Handle author name prefixes and check bibliography
Inspected pdf output of bibliography and fixed minor things.

Also added some biblatex code to correctly handle prefixes such as
"van" or "de". They are now not capitalized and sort correctly in the
bibliography (e.g. "de Boer" is with B and not D).

Note that this breaks sorting/capitalization in html, but that is less
important than PDF. Another advantage is that capitalization doesn't
have to be changed manually in the .bib anymore after exporting
(prefixes have to be capitalized to get Mendeley to recognize them as
part of the surname; but now biblatex handles everything).